Migrate purchase token verification to subscriber schema validation
